pyproject.toml
[project]
name = "face-recognition-service"
version = "1.0.0"
description = "Enterprise-grade 1:N face recognition service"
authors = [{name = "ML Platform Team", email = "ml-platform@company.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.25.0",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",
    "sqlalchemy[asyncio]==2.0.25",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "insightface==0.7.3",
    "onnxruntime==1.16.3",
    "opencv-python-headless==4.9.0.80",
    "faiss-cpu==1.7.4",
    "numpy==1.24.3",
    "pillow==10.2.0",
    "structlog==24.1.0",
    "prometheus-client==0.19.0",
    "python-multipart==0.0.6",
    "httpx==0.26.0",
    "tenacity==8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "ruff==0.1.11",
    "black==23.12.1",
    "isort==5.13.2",
    "pre-commit==3.6.0",
    "faker==22.0.0",
]
gpu = [
    "faiss-gpu==1.7.4",
    "onnxruntime-gpu==1.16.3",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "UP", "S", "B", "A", "C4", "DTZ", "T20", "SIM"]
ignore = ["E501", "S101"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"